requirements.txt
requests==2.24.0
SQLAlchemy==1.3.19
pytest==6.0.1
Flask_SQLAlchemy==2.4.4
Flask_RESTful==0.3.8
Flask_APScheduler==1.11.0
Flask==1.1.2
python-dotenv==0.14.0
